Split ContentView.swift file by views
